添加模型加载和光照项目

+++ includes/learnopengl/mesh.h
@@ -69,13 +69,13 @@ public:
             string number;
             string name = textures[i].type;
             if(name == "texture_diffuse")
-				number = std::to_string(diffuseNr++);
+				number = std::to_string((long long)diffuseNr++);
 			else if(name == "texture_specular")
-				number = std::to_string(specularNr++); // transfer unsigned int to stream
+				number = std::to_string((long long)specularNr++); // transfer unsigned int to stream
             else if(name == "texture_normal")
-				number = std::to_string(normalNr++); // transfer unsigned int to stream
+				number = std::to_string((long long)normalNr++); // transfer unsigned int to stream
              else if(name == "texture_height")
-			    number = std::to_string(heightNr++); // transfer unsigned int to stream
+			    number = std::to_string((long long)heightNr++); // transfer unsigned int to stream
 
 													 // now set the sampler to the correct texture unit
             glUniform1i(glGetUniformLocation(shader.ID, (name + number).c_str()), i);
